backend/requirements.txt
aiocontextvars
aiofiles
asgiref
asn1crypto
asyncpg
cffi
ciso8601
Click
contextvars
cryptography
fastadmin[fastapi]
gunicorn
h11
httptools
immutables
itsdangerous
Jinja2
marshmallow
MarkupSafe
py-bcrypt
pycparser
PyJWT
PyPika
python-multipart
python-dotenv
pytz
secure
six
starlette
tortoise-orm<=0.20.1
uvicorn
uvloop
webargs
WTForms
